rewriting to unique pointer and main to framework

+++ namiEngine/namiEngine/utility/Angle.cpp
@@ -7,7 +7,7 @@ void Angle::Update() {
 	angle1 = endPos1 - startPos1;
 	angle2 = endPos2 - startPos2;
 
-	resultRadians = atan2(angle1, angle2);
+	resultRadians = static_cast<float>(atan2(angle1, angle2));
 }
 
 float Angle::GetRadians() {
@@ -18,7 +18,7 @@ float Angle::PosForRadians(float startPos1, float startPos2, float endPos1, floa
 	angle1 = endPos1 - startPos1;
 	angle2 = endPos2 - startPos2;
 
-	resultRadians = atan2(angle1, angle2);
+	resultRadians = static_cast<float>(atan2(angle1, angle2));
 
 	return resultRadians;
 }
@@ -27,7 +27,7 @@ float Angle::lR(float startPos1, float startPos2, float endPos1, float endPos2) 
 	float px = endPos1 - startPos1;
 	float py = endPos2 - startPos2;
 
-	float l = sqrt(px * px + py * py);
+	float l = static_cast<float>(sqrt(px * px + py * py));
 	return l;
 }
